simple-cam: Move camera detection check to where it belongs
Commit 14529b6d1c4a ("simple-cam: Early return if no cameras are found
on the system") added a check to ensure that at least one camera is
present in the system to avoid a crash otherwise. The check was however
placed towards the end of the main() function, way after the camera is
accessed, possibly due to a rebase conflict. Move it before accessing
the first camera.

While at it, replace usage of std::vector::size() with
std::vector::empty() which expresses the intent better.

Fixes: 14529b6d1c4a ("simple-cam: Early return if no cameras are found on the system")
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>