simple-cam: processRequest: Report Request metadata
The completed Request can contain extra data associated with the
completed capture to report IPA state or properties of the capture to
the application.

Process and report any metadata that is associated with the request
when it completes, to demonstrate the extra metadata available to be
processed.

Now that there is substantially more information than one line per
completed request, start each completed Request with a blank line to
separate from previous completions and report the state of the Request
itself with its .toString() helper.

Suggested-by: Nejc Galof <galof.nejc@gmail.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>